Cube new

+++ src/games/TStats.java
@@ -82,7 +82,7 @@ public class TStats {
 				    				// scrambled cube may be created with p twist, it may happen, that it belongs
 				    				// to distance set D[p-1] or lower and can be solved with p-1 twists or less.
 				    if (cs.p<cs.moveNum && cs.moveNum<cs.epiLength) nLonger++;	// completed in p+1,...,epiLength-1 moves
-				    if (cs.moveNum>=cs.epiLength) nNot++;	// did not complete after epiLength moves		    	
+				    if (cs.moveNum>=cs.epiLength) nNot++;	// did not complete after epiLength moves
 			    }
 	        } 
 		    this.size = size;
@@ -129,7 +129,7 @@ public class TStats {
 	 * @param mode =0: percent solved within minimal twists, =1: percent solved below epiLength
 	 * @return the weighted average of 'solved' percentages in {@code taList}
 	 */
-	public static double weightedAvgResTAggregList(ArrayList<TAggreg> taList, int[] w, int mode) {
+	public static double weightedAvgResTAggregList(ArrayList<TAggreg> taList, double[] w, int mode) {
 		assert (w.length >= taList.size());
 		Iterator it = taList.iterator();
 		double res=0;
